fix logo not display & remove unused files
