--- client/features/home/CalendarPage.js
@@ -1,42 +1,0 @@
-import { useAuthContext } from "client/contexts/AuthContext";
-import FormProvider from "client/contexts/FormContext";
-import { useFormModalContext } from "client/contexts/FormModalContext";
-import { useModalContext } from "client/contexts/ModalContext";
-import UserForm from "client/features/form/UserForm";
-import EventModal from "client/features/modal/EventModal";
-import Modal from "client/features/modal/Modal";
-import useDate from "client/hooks/useDate";
-
-import Calendar from "../calendar/Calendar";
-import CalendarHeader from "../calendarHeader";
-
-function CalendarPage() {
-  const auth = useAuthContext();
-  const date = useDate();
-  const formModal = useFormModalContext();
-  const modal = useModalContext();
-
-  return (
-    <>
-      <main className="flex flex-col gap-3 p-3 shadow-sm min-h-screen max-w-[1920px] mx-auto">
-        <CalendarHeader date={date} />
-        <Calendar date={date} />
-      </main>
-      {auth?.user && (
-        <FormProvider>
-          <div className="md:w-1/2 mx-auto shadow-xl rounded-2xl pb-2 bg-white">
-            <Modal context={modal}>
-              <EventModal />
-            </Modal>
-            {auth?.user && 
-            <Modal context={formModal}>
-              <UserForm />
-            </Modal>
-            }
-          </div>
-        </FormProvider>
-      )}
-    </>
-  );
-}
-export default CalendarPage;
