+++ client/features/modal/Modal.jsx
@@ -1,6 +1,6 @@
-import { AnimatePresence } from "framer-motion";
+import { AnimatePresence , motion } from "framer-motion";
+
 import Backdrop from "./Backdrop";
-import { motion } from "framer-motion";
 
 const Modal = ({ children, context }) => {
